requirements.txt
langchain==0.1.0
langchain-community
pypdf
sentence-transformers
faiss-cpu
torch
transformers
streamlit
numpy